modern_robotics

Module dynamics_of_open_chains

Functions

ad

Calculates the 6x6 matrix [adV] of a spatial velocity vector.

end_effector_forces

Computes the joint forces/torques required to create the end-effector force Ftip.

euler_step

Computes the joint angles and velocities at the next timestep using first order Euler integration.

forward_dynamics

Computes forward dynamics in the space frame for an open chain robot.

forward_dynamics_trajectory

Simulates the motion of a serial chain given an open-loop history of joint forces/torques.

gravity_forces

Computes the joint forces/torques required to overcome gravity for an open chain robot.

inverse_dynamics

Computes inverse dynamics in the space frame for an open chain robot.

inverse_dynamics_trajectory

Calculates the required joint forces/torques to achieve a path with given accelerations.

mass_matrix

Computes the mass matrix of an open chain robot.

vel_quadratic_forces

Computes the Coriolis and centripetal terms in the inverse dynamics of an open chain robot.