modern_robotics::dynamics_of_open_chains

Function inverse_dynamics_trajectory

pub fn inverse_dynamics_trajectory(
    thetamat: &Vec<DVector<f64>>,
    dthetamat: &Vec<DVector<f64>>,
    ddthetamat: &Vec<DVector<f64>>,
    g: &Vector3<f64>,
    ftipmat: &Vec<Vector6<f64>>,
    mlist: &Vec<Matrix4<f64>>,
    glist: &Vec<Matrix6<f64>>,
    slist: &Vec<Vector6<f64>>,
) -> Vec<DVector<f64>> 

Calculates the required joint forces/torques to achieve a path with given accelerations.

Arguments

thetamat - A trajectory of joint variables, each row is a point in the trajectory

dthetamat - A trajectory of joint velocities

ddthetamat - A trajectory of joint accelerations

g - Gravity vector g

ftipmat - A trajectory of spatial forces applied by the end-effector (i-th element is the spatial force applied by the end-effector at the i-th time step)

mlist - List of link frames {i} relative to {i-1} at the home position

glist - Spatial inertia matrices Gi of the links

slist - Screw axes Si of the joints in a space frame

Returns

The N x n matrix of joint forces/torques for the specified trajectory, where each of the N rows is the vector of joint forces/torques at each time step

This function uses inverse_dynamics to calculate the joint forces/torques required to move along the given trajectory.

Example

use nalgebra::{Vector3, Vector6, DVector, Matrix4, Matrix6};
use modern_robotics::inverse_dynamics_trajectory;

let thetamat = vec![DVector::from_vec(vec![0.1, 0.1, 0.1])];
let dthetamat = vec![DVector::from_vec(vec![0.1, 0.2, 0.3])];
let ddthetamat = vec![DVector::from_vec(vec![2.0, 1.5, 1.0])];
let g = Vector3::new(0.0, 0.0, -9.8);
let ftipmat = vec![Vector6::new(1.0, 1.0, 1.0, 1.0, 1.0, 1.0)];
// ... setup mlist, glist, slist ...
let taumat = inverse_dynamics_trajectory(&thetamat, &dthetamat, &ddthetamat,
                                          &g, &ftipmat, &mlist, &glist, &slist);