modern_robotics

Function forward_dynamics

pub fn forward_dynamics(
    thetalist: &DVector<f64>,
    dthetalist: &DVector<f64>,
    taulist: &DVector<f64>,
    g: &Vector3<f64>,
    ftip: &Vector6<f64>,
    mlist: &Vec<Matrix4<f64>>,
    glist: &Vec<Matrix6<f64>>,
    slist: &Vec<Vector6<f64>>,
) -> DVector<f64>

Computes forward dynamics in the space frame for an open chain robot.

Arguments

thetalist - A list of joint variables

dthetalist - A list of joint rates

taulist - An n-vector of joint forces/torques

g - Gravity vector g

ftip - Spatial force applied by the end-effector expressed in frame {n+1}

mlist - List of link frames {i} relative to {i-1} at the home position

glist - Spatial inertia matrices Gi of the links

slist - Screw axes Si of the joints in a space frame

Returns

The resulting joint accelerations

This function computes ddthetalist by solving: M(thetalist) * ddthetalist = taulist - c(thetalist,dthetalist) - g(thetalist) - Jtr(thetalist) * Ftip

Example

use nalgebra::{Vector3, Vector6, DVector, Matrix4, Matrix6};
use modern_robotics::forward_dynamics;

let thetalist = DVector::from_vec(vec![0.1, 0.1, 0.1]);
let dthetalist = DVector::from_vec(vec![0.1, 0.2, 0.3]);
let taulist = DVector::from_vec(vec![0.5, 0.6, 0.7]);
let g = Vector3::new(0.0, 0.0, -9.8);
let ftip = Vector6::zeros();
// ... setup mlist, glist, slist ...
let ddthetalist = forward_dynamics(&thetalist, &dthetalist, &taulist,
                                    &g, &ftip, &mlist, &glist, &slist);