modern_robotics

Function forward_dynamics_trajectory

pub fn forward_dynamics_trajectory(
    thetalist: &DVector<f64>,
    dthetalist: &DVector<f64>,
    taumat: &Vec<DVector<f64>>,
    g: &Vector3<f64>,
    ftipmat: &Vec<Vector6<f64>>,
    mlist: &Vec<Matrix4<f64>>,
    glist: &Vec<Matrix6<f64>>,
    slist: &Vec<Vector6<f64>>,
    dt: f64,
    int_res: i32,
) -> (Vec<DVector<f64>>, Vec<DVector<f64>>)

Simulates the motion of a serial chain given an open-loop history of joint forces/torques.

Arguments

thetalist - n-vector of initial joint variables

dthetalist - n-vector of initial joint rates

taumat - An N x n matrix of joint forces/torques, where each row is the joint effort at a point in the trajectory

g - Gravity vector g

ftipmat - An N x 6 matrix of spatial forces applied by the end-effector (i-th row is the spatial force applied at the i-th time step)

mlist - List of link frames {i} relative to {i-1} at the home position

glist - Spatial inertia matrices Gi of the links

slist - Screw axes Si of the joints in a space frame

dt - The timestep between consecutive joint forces/torques

int_res - Integration resolution: the number of Euler steps during each timestep (typically set to 1-10)

Returns

A tuple containing:

The resulting joint angle trajectory

The resulting joint velocity trajectory

Uses forward dynamics and Euler integration.

Example

use nalgebra::{Vector3, Vector6, DVector, Matrix4, Matrix6};
use modern_robotics::forward_dynamics_trajectory;

let thetalist = DVector::from_vec(vec![0.1, 0.1, 0.1]);
let dthetalist = DVector::from_vec(vec![0.1, 0.2, 0.3]);
let taumat = vec![DVector::from_vec(vec![0.1, 0.1, 0.1])];
let g = Vector3::new(0.0, 0.0, -9.8);
let ftipmat = vec![Vector6::zeros()];
// ... setup mlist, glist, slist ...
let dt = 0.01;
let int_res = 8;
let (thetamat, dthetamat) = forward_dynamics_trajectory(&thetalist, &dthetalist,
                                                         &taumat, &g, &ftipmat,
                                                         &mlist, &glist, &slist,
                                                         dt, int_res);