modern_robotics

Function joint_trajectory

pub fn joint_trajectory(
    thetastart: &DVector<f64>,
    thetaend: &DVector<f64>,
    tf: f64,
    n: usize,
    method: Method,
) -> Vec<DVector<f64>> 

Computes a straight-line trajectory in joint space.

Arguments

thetastart - The initial joint variables

thetaend - The final joint variables

tf - Total time of the motion in seconds from rest to rest

n - The number of points N > 1 (Start and stop) in the discrete representation of the trajectory

method - The time-scaling method, where Method::Cubic uses a third-order polynomial and Method::Quintic uses a fifth-order polynomial

Returns

A trajectory as a list of N points, where each point is a vector of joint coordinates. The first point is thetastart and the Nth point is thetaend.

This function is similar to the MATLAB/Octave function jtraj.

Example

use nalgebra::DVector;
use modern_robotics::{joint_trajectory, Method};

let thetastart = DVector::from_vec(vec![1.0, 0.0, 0.0, 1.0, 1.5, 2.5, 3.0, 0.0]);
let thetaend = DVector::from_vec(vec![1.2, 0.5, 0.6, 1.1, 2.0, 3.1, 3.2, 0.9]);
let tf = 4.0;
let n = 6;
let method = Method::Cubic;
let traj = joint_trajectory(&thetastart, &thetaend, tf, n, method);