modern_robotics

Module rigid_body_motions

Functions

adjoint

Computes the adjoint representation of a homogeneous transformation matrix

axis_ang3

Converts a 3-vector of exponential coordinates for rotation into axis-angle form

axis_ang6

Converts a 6-vector of exponential coordinates into screw axis-angleform

matrix_exp3

Computes the matrix exponential of a matrix in so(3)

matrix_exp6

Computes the matrix exponential of an se3 representation of exponential coordinates

matrix_log3

Computes the matrix logarithm of a rotation matrix

matrix_log6

Computes the matrix logarithm of a homogeneous transformation matrix

rot_inv

Inverts a rotation matrix.

rp_to_trans

Converts a rotation matrix and a position vector into homogeneous transformation matrix

screw_to_axis

Takes a parametric description of a screw axis and converts it to a normalized screw axis.

se3_to_vec

Converts an se3 matrix into a spatial velocity vector

so3_to_vec

Converts an so(3) representation to a 3-vector

trans_inv

Inverts a homogeneous transformation matrix

trans_to_rp

Converts a homogeneous transformation matrix into a rotation matrix and position vector

vec_to_se3

Converts a spatial velocity vector into a 4x4 matrix in se3

vec_to_so3

Converts a 3-vector to an so(3) representation