modern_robotics

Module trajectory_generation

Enums

Method

Time scaling method for trajectory generation.

Functions

cartesian_trajectory

Computes a trajectory as a list of N SE(3) matrices with Cartesian motion.

cubic_time_scaling

Computes s(t) for a cubic time scaling.

joint_trajectory

Computes a straight-line trajectory in joint space.

quintic_time_scaling

Computes s(t) for a quintic time scaling.

screw_trajectory

Computes a trajectory as a list of N SE(3) matrices corresponding to the screw motion.