modern_robotics::trajectory_generation

Function screw_trajectory

pub fn screw_trajectory(
    xstart: &Matrix4<f64>,
    xend: &Matrix4<f64>,
    tf: f64,
    n: usize,
    method: Method,
) -> Vec<Matrix4<f64>> 

Computes a trajectory as a list of N SE(3) matrices corresponding to the screw motion.

Arguments

xstart - The initial end-effector configuration

xend - The final end-effector configuration

tf - Total time of the motion in seconds from rest to rest

n - The number of points N > 1 (Start and stop) in the discrete representation of the trajectory

method - The time-scaling method, where Method::Cubic uses a third-order polynomial and Method::Quintic uses a fifth-order polynomial

Returns

The discretized trajectory as a list of N matrices in SE(3) separated in time by Tf/(N-1). The first point is Xstart and the Nth point is Xend.

This function calculates a trajectory corresponding to the screw motion about a space screw axis.

Example

use nalgebra::Matrix4;
use modern_robotics::{screw_trajectory, Method};

let xstart = Matrix4::new(
    1.0, 0.0, 0.0, 1.0,
    0.0, 1.0, 0.0, 0.0,
    0.0, 0.0, 1.0, 1.0,
    0.0, 0.0, 0.0, 1.0
);
let xend = Matrix4::new(
    0.0, 0.0, 1.0, 0.1,
    1.0, 0.0, 0.0, 0.0,
    0.0, 1.0, 0.0, 4.1,
    0.0, 0.0, 0.0, 1.0
);
let tf = 5.0;
let n = 4;
let method = Method::Cubic;
let traj = screw_trajectory(&xstart, &xend, tf, n, method);